Own the backtracker's per-match state
`backpos` was the last garray in the regexp family and the reason
`reg_save`/`reg_restore` — the two functions in the innermost loop of the
backtracking engine — took a `*mut garray_T` from twenty sites and used it
only for `ga_len`.

It becomes `BackPos`, a `Vec` beside `RegStack` and the `\{n,m}` counters in
one `BtState` that `regmatch` takes once and threads down by `&mut`.
`reg_save` now *returns* a `SavedInput` instead of writing through an
out-parameter, which is what dissolves F-P21-8b: the `BACK` arm used to hold
a raw pointer **into** the array (`seen().add(i)`) across a `reg_save` call,
and it now asks `BackPos::stepped` whether the loop moved and never forms an
interior reference at all. A back edge only ever stored a position, so its
`backpos_len` half was dead weight; `BackEdge` drops it.

`brace_min`/`brace_max`/`brace_count` were three parallel ten-slot globals
indexed by the same operand at every use; they are one `Brace` per slot now.

## The buffer is longer than the record, deliberately

`BackPos` keeps `seen` (everything written this match) and `live` (how many
the record holds) rather than truncating, because upstream's `ga_len` is a
bare cursor and a restore can put it back *up*: a look-behind restores to
`save_after` after an attempt that discovered fewer edges than were live when
it was saved, and upstream then reads the entries past the cursor with
whatever they last held. Truncating a `Vec` would silently drop them and
change which patterns match.

## Why it is taken raw

`with_mut` at the two entry points cost **3.7×** on a regexp-heavy debug
workload — 2.7 s → 10.2 s — because `regmatch` runs once per start column and
the debug borrow table is a thread-local lookup plus a hash-map insert each
time. That is F-P21-8c measured. The state is taken with `.ptr()` once per
match, the way `Rex::acquire` is, and threaded from there; after the rewrite
the same workload is 2.69–2.76 s against a 2.70–2.76 s baseline.

`cell_ptr` 1,400 -> 1,362. `resweep` IDENTICAL (txt/err/stderr); `cargo test`
701 + 662; `functionaltest test/functional/editor` 534 passed 0 failed;
`unittest` green; `oldtest test_regexp_utf8 test_regexp_latin test_substitute
test_search` 197 tests, 0 failed.

+++ crates/nvim/src/regexp/pos.rs
@@ -46,8 +46,6 @@
     clippy::ptr_as_ptr
 )]
 
-use core::ffi::c_int;
-
 use crate::types::{lpos_T, uint8_t};
 
 /// Which shape a match records positions in. Fixed for the whole of one
@@ -247,7 +245,7 @@ impl Capture {
 #[derive(Clone, Copy)]
 pub(crate) struct SavedInput {
     pub(crate) pos: MatchPos,
-    pub(crate) backpos_len: c_int,
+    pub(crate) backpos_len: usize,
 }
 
 impl SavedInput {
